Add alert for befor delete any post or comment

+++ public/assets/js/posts.js
@@ -231,7 +231,23 @@ window.onload = () => {
   // -------------------- function to delete post ----------------------
   const deletePost = (postId) => {
     if (userID) {
-      deletePostUser({ postId })
+      useConfirmAlert(
+        'Are you sure?',
+        "You won't Delete this Post!",
+        'warning',
+        true,
+        '#3085d6',
+        '#e55f34',
+        'Yes, delete it!',
+      )
+        .then((result) => {
+          if (result.isConfirmed) {
+            return deletePostUser({ postId });
+          }
+          const cancel = new Error('Cancel delete post');
+          cancel.type = 'cancel';
+          throw cancel;
+        })
         .then(({ status, message }) => {
           if (status === 200) {
             lastFivePostAdded();
@@ -257,16 +273,18 @@ window.onload = () => {
             removeDefaultPost();
           }
         })
-        .catch((error) => {
-          useAlert(
-            'Error!',
-            'Sorry! Some things went wrong',
-            'error',
-            'Ok',
-            'center',
-            2000,
-            false,
-          );
+        .catch(({ type }) => {
+          if (type !== 'cancel') {
+            useAlert(
+              'Error!',
+              'Sorry! Some things went wrong',
+              'error',
+              'Ok',
+              'center',
+              2000,
+              false,
+            );
+          }
         });
     } else {
       window.location.href = '/auth/login';
